Fix naming mismatch
Project names did not match the folders in which they were contained
