+++ StrategyPattern.After/Domain/Order.cs
@@ -1,4 +1,4 @@
-namespace StrategyPattern.Before.Domain
+namespace StrategyPattern.After
 {
     public class Order
     {
