+++ StrategyPattern.After/Domain/ShippingService/IShippingCostStrategy.cs
