+++ StrategyPattern.After/StrategyPattern.After.csproj
