+++ StrategyPattern.Before/Domain/Address.cs
@@ -1,4 +1,4 @@
-namespace StrategyPattern.After
+namespace StrategyPattern.Before.Domain
 {
     public class Address
     {
