+++ StrategyPattern.Before/StrategyPattern.Before.csproj
